:truck: Restructured files and APIs to align with latest libhal
