+++ v5/include/libhal-util/usb/utils.hpp
@@ -1,12 +1,27 @@
+// Copyright 2024 - 2025 Khalil Estell and the libhal contributors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 #pragma once
 
-#include <cstddef>
+#include <libhal/error.hpp>
 #include <libhal/units.hpp>
 #include <libhal/usb.hpp>
 
-namespace hal::usb {
+namespace hal::v5::usb {
 
 namespace constants {
+
 constexpr byte device_desc_size = 18;
 constexpr byte config_desc_size = 9;
 constexpr byte inferface_desc_size = 9;
@@ -14,12 +29,11 @@ constexpr byte endpoint_desc_size = 7;
 constexpr byte iad_desc_size = 0x08;
 
 constexpr byte size_std_req = 8;
+
 }  // namespace constants
 
-/**
- * @brief USB Class Code indicating the type of device or interface. Assigned by
- * USB-IF.
- */
+// Maybe move these enum classes into the constants namespace
+// Assigned by USB-IF
 enum class class_code : hal::byte
 {
   use_interface_descriptor =
@@ -50,34 +64,35 @@ enum class class_code : hal::byte
   vendor_specific = 0xFF        // Vendor Specific
 };
 
-/**
- * @brief USB Descriptor Type values as defined by the USB specification.
- * These values are used in the bDescriptorType field of USB descriptors
- * to identify the type and structure of the descriptor data.
- */
+// Default types
 enum class descriptor_type : hal::byte
 {
-  device = 0x1,                        // Device descriptor
-  configuration = 0x2,                 // Configuration descriptor
-  string = 0x3,                        // String descriptor
-  interface = 0x4,                     // Interface descriptor
-  endpoint = 0x5,                      // Endpoint descriptor
-  device_qualifier = 0x6,              // Device qualifier descriptor
-  other_speed_configuration = 0x7,     // Other speed configuration descriptor
-  interface_power = 0x8,               // Interface power descriptor
-  otg = 0x9,                           // OTG descriptor
-  debug = 0xA,                         // Debug descriptor
-  interface_association = 0xB,         // Interface Association descriptor
-  security = 0xC,                      // Security descriptor
-  key = 0xD,                           // Key descriptor
-  encryption_type = 0xE,               // Encryption type descriptor
-  bos = 0xF,                           // Binary Object Store (BOS) descriptor
-  device_capability = 0x10,            // Device capability descriptor
-  wireless_endpoint_companion = 0x11,  // Wireless endpoint companion descriptor
-  superspeed_endpoint_companion =
-    0x30,  // SuperSpeed endpoint companion descriptor
-  superspeed_endpoint_isochronous_companion =
-    0x31,  // SuperSpeed isochronous endpoint companion descriptor
+  device = 0x1,
+  configuration = 0x2,
+  string = 0x3,
+  interface = 0x4,
+  endpoint = 0x5,
+  device_qualifier = 0x6,
+  other_speed_configuration = 0x7,
+  interface_power = 0x8,
+  otg = 0x9,
+  debug = 0xA,
+  interface_association = 0xB,
+  security = 0xC,
+  key = 0xD,
+  encryption_type = 0xE,
+  bos = 0xF,
+  device_capability = 0x10,
+  wireless_endpoint_companion = 0x11,
+  superspeed_endpoint_companion = 0x30,
+  superspeed_endpoint_isochronous_companion = 0x31,
 };
+
+}  // namespace hal::v5::usb
+
+namespace hal::usb {
+namespace constants = v5::usb::constants;
+using v5::usb::class_code;
+using v5::usb::descriptor_type;
 
 }  // namespace hal::usb
